[Sigma BugID #42200] Sigma DTV: add support for armv8 (init for union)
1. initial uboot for union
   with peripheral drivers excluded
2. validated on veloce2 (make union_veloce_defconfig)

+++ arch/arm/include/asm/arch-trix/bootparam.h
@@ -46,7 +46,7 @@
  * This structure defines boot param for uboot
  */
 struct bootparam_frame{
-	long data[(BOOTPARAM_FRM_SIZE) >> 2];
+	u32 data[(BOOTPARAM_FRM_SIZE) >> 2];
 };
 
 #define BOOTPARAM_TAIL	data[BOOTPARAM_TAIL_OFS >> 2]
@@ -59,9 +59,9 @@ struct bootparam_frame{
 
 extern struct bootparam_frame boot_params;
 char* sx6_boot_message(void);
-long sx6_boot_device(void);
-long sx6_boot_reason(void);
-long sx6_boot_flags(void);
+u32 sx6_boot_device(void);
+u32 sx6_boot_reason(void);
+u32 sx6_boot_flags(void);
 #endif //__ASSEMBLY__
 
 #endif
